Got everything mostly working now, only issue I can see is that it saves multiple autosaves

+++ Assets/Scripts/GameFramework/Services/Data/TimeStatistics.cs
@@ -7,7 +7,7 @@ namespace GameFramework.Services.Data
     /// </summary>
     public struct TimeStatistics
     {
-        public float GameTime;
+        public double GameTime;
         public float SessionTime;
         public float LevelTime;
         public bool IsTrackingGameTime;
